refactor(react, react-query): remove __test__ directory to collocate
