+++ configs/tsup/src/index.ts
@@ -3,7 +3,12 @@ import { Options } from 'tsup'
 export const options: Options = {
   banner: { js: '"use client"' },
   format: ['cjs', 'esm'],
-  entry: ['src/*.ts', 'src/*.tsx', 'src/experimental/*.ts', 'src/experimental/*.tsx'],
+  entry: [
+    'src/*.{ts,tsx}',
+    '!src/*.{spec,test,test-d}.*',
+    'src/experimental/*.{ts,tsx}',
+    '!src/experimental/*.{spec,test,test-d}.*',
+  ],
   sourcemap: true,
   dts: true,
   splitting: false,
