+++ packages/react/src/AsyncBoundary.spec.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react'
-import { AsyncBoundary } from '..'
-import { FALLBACK, TEXT } from './utils'
+import { FALLBACK, TEXT } from './utils.spec'
+import { AsyncBoundary } from '.'
 
 describe('AsyncBoundary', () => {
   it('should show children', () => {
