+++ packages/react/src/ErrorBoundary.spec.tsx
@@ -1,8 +1,8 @@
 import { act } from '@testing-library/react'
 import { ComponentProps, ComponentRef, createRef } from 'react'
 import { createRoot } from 'react-dom/client'
-import { ErrorBoundary } from '..'
-import { ERROR_MESSAGE, FALLBACK, MS_100, TEXT, ThrowError } from './utils'
+import { ERROR_MESSAGE, FALLBACK, MS_100, TEXT, ThrowError } from './utils.spec'
+import { ErrorBoundary } from '.'
 
 let container = document.createElement('div')
 let root = createRoot(container)
