+++ packages/react/src/ErrorBoundaryGroup.spec.tsx
@@ -1,6 +1,6 @@
 import { act, render, screen } from '@testing-library/react'
-import { ErrorBoundary, ErrorBoundaryGroup, useErrorBoundaryGroup } from '..'
-import { ERROR_MESSAGE, MS_100, TEXT, ThrowError } from './utils'
+import { ERROR_MESSAGE, MS_100, TEXT, ThrowError } from './utils.spec'
+import { ErrorBoundary, ErrorBoundaryGroup, useErrorBoundaryGroup } from '.'
 
 const innerErrorBoundaryCount = 3
 const resetButtonText = 'reset button'
