+++ packages/react/src/experimental/Await.spec.tsx
@@ -1,7 +1,7 @@
 import { act, render, screen, waitFor } from '@testing-library/react'
-import { ErrorBoundary, Suspense } from '../..'
-import { suspensiveCache, useAwait } from '../../experimental'
-import { ERROR_MESSAGE, FALLBACK, MS_100, TEXT, delay } from '../utils'
+import { ErrorBoundary, Suspense } from '..'
+import { ERROR_MESSAGE, FALLBACK, MS_100, TEXT, delay } from '../utils.spec'
+import { suspensiveCache, useAwait } from '.'
 
 const cacheFnSuccess = () => delay(MS_100).then(() => TEXT)
 const cacheFnFailure = () =>
