+++ packages/react/src/withSuspense.spec.tsx
@@ -1,6 +1,6 @@
 import { render, screen, waitFor } from '@testing-library/react'
-import { withSuspense } from '..'
-import { FALLBACK, MS_100, Suspend, TEXT } from './utils'
+import { FALLBACK, MS_100, Suspend, TEXT } from './utils.spec'
+import { withSuspense } from '.'
 
 const SuspendDuring100msToShowTEXTInSuspense = withSuspense(() => <Suspend during={MS_100} toShow={TEXT} />, {
   fallback: FALLBACK,
